+++ features/mpaccount/src/main/java/com/jpp/mpaccount/login/LoginViewState.kt
@@ -1,0 +1,40 @@
+package com.jpp.mpaccount.login
+
+import android.view.View
+import com.jpp.mpdesign.views.MPErrorView.ErrorViewState
+
+/**
+ * Represents the view state of the login screen.
+ */
+data class LoginViewState(
+    val loadingVisibility: Int = View.INVISIBLE,
+    val errorViewState: ErrorViewState = ErrorViewState.asNotVisible(),
+    val oauthViewState: OauthViewState = OauthViewState()
+) {
+    companion object {
+        fun showLoading() = LoginViewState(loadingVisibility = View.VISIBLE)
+        fun showNoConnectivityError(errorHandler: () -> Unit) = LoginViewState(
+            errorViewState = ErrorViewState.asConnectivity(errorHandler)
+        )
+
+        fun showUnknownError(errorHandler: () -> Unit) = LoginViewState(
+            errorViewState = ErrorViewState.asUnknownError(errorHandler)
+        )
+
+        fun showOauth(
+            url: String,
+            interceptUrl: String,
+            reminder: Boolean,
+            redirectListener: (String) -> Unit
+        ) = LoginViewState(
+            oauthViewState = OauthViewState(
+                visibility = View.VISIBLE,
+                url = url,
+                interceptUrl = interceptUrl,
+                reminder = reminder,
+                redirectListener = redirectListener
+            )
+        )
+    }
+}
+
