+++ features/mpaccount/src/main/java/com/jpp/mpaccount/login/OauthViewState.kt
@@ -1,0 +1,14 @@
+package com.jpp.mpaccount.login
+
+import android.view.View
+
+/**
+ * Represents the view state of the Oauth section of the login screen.
+ */
+data class OauthViewState(
+    val visibility: Int = View.INVISIBLE,
+    val url: String? = null,
+    val interceptUrl: String? = null,
+    val redirectListener: ((String) -> Unit)? = null,
+    val reminder: Boolean = false
+)
